requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
python-dotenv==0.18.0
pandas==1.3.2
numpy==1.21.2
pytesseract==0.3.8
pillow==8.3.1
pymupdf==1.18.19
reportlab==3.6.1
matplotlib==3.4.3
scikit-learn==1.0.2
